+++ Scripts/Game/DCO/Arsenal/Vehicle/DCO_VehicleServiceMenu.c
@@ -51,6 +51,8 @@ class DCO_VehicleServiceButtonHandler : ScriptedWidgetEventHandler
 
 	override bool OnClick(Widget w, int x, int y, int button)
 	{
+		if (button != 0)
+			return false;
 		return m_Menu && m_Menu.OnAction(m_iAction);
 	}
 }