+++ Scripts/Game/DCO/GMUI/Create/DCO_GMCreatePanelComponent.c
@@ -11,6 +11,8 @@ class DCO_CreatePanelButtonHandler : ScriptedWidgetEventHandler
 
 	override bool OnClick(Widget w, int x, int y, int button)
 	{
+		if (button != 0)
+			return false;
 		if (m_Owner)
 			return m_Owner.OnButton(w);
 		return false;