+++ Scripts/Game/DCO/GMUI/Orders/DCO_GMOrdersPanel.c
@@ -10,6 +10,8 @@ class DCO_OrdersPanelHandler : ScriptedWidgetEventHandler
 
 	override bool OnClick(Widget w, int x, int y, int button)
 	{
+		if (button != 0)
+			return false;
 		if (m_Owner)
 			return m_Owner.OnButton(w);
 		return false;